% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose_trvaltest_up_to_year.R
\name{compose_trvaltest_up_to_year}
\alias{compose_trvaltest_up_to_year}
\title{Compose dataset for training/test}
\usage{
compose_trvaltest_up_to_year(db, year)
}
\arguments{
\item{db}{(data frame) original dataset, must have a variable \code{anno}}

\item{year}{(int) integerish number form 2004 to 2013 (we have data
up to 2014) so we cannot have more than 2013 data if we want to
have data to predict}
}
\value{

}
\description{
Starting from the dataset provided, it creates the
train/validation/test dataset consideraing training data up to
\code{year}. Test set wil be two years forward \code{year}.
}
