% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_varicella_mixdb.R
\name{create_varicella_mixdb}
\alias{create_varicella_mixdb}
\title{Create the main mixdb for the vericella project}
\usage{
create_varicella_mixdb(gold_varicella)
}
\arguments{
\item{gold_varicella}{(data frame) full varicella data to use,
including a variable named \code{set} (with tag for "train" and
"validation", and possibly "test" too), and a factor variable named
\code{class} (with the class of interest, e.g. "negative" and
"positive")}
}
\value{
invisibly `TRUE``
}
\description{
Create the main \link{mixdb} for the varicella project
}
