% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_dates_for_gold.R
\name{eval_dates_for_gold}
\alias{eval_dates_for_gold}
\title{Create dated gold standard}
\usage{
eval_dates_for_gold(gold, varic_df)
}
\arguments{
\item{gold}{(df) the gold standard}

\item{varic_df}{the whole varicella dataset}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Create a version of the gold standard with the first date in which
a patient became positive to Zoster
}
\examples{
\dontrun{
  # inside the development project
  library(targets)

  gold <- tar_read(gold)
  df <- tar_read(varic_df)

  eval_dates_for_gold(gold, varic_df)
}

}
