% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_prepost_positive.R
\name{expand_prepost_positive}
\alias{expand_prepost_positive}
\title{Expand previous and post positive status}
\usage{
expand_prepost_positive(db)
}
\arguments{
\item{db}{(data frame) positive database with \code{Id_Medico}, \code{nPaz},
and \code{Data} as the minimum set of columns. Column \code{Data} reporting the
year of the Varicella infection for the corresponding patient.}
}
\value{
the original \link[tibble:tibble-package]{tibble} with the
\code{is_was_positive} column added as last one.
}
\description{
Given a dataset with columns \code{Id_Medico}, \code{nPaz}, and \code{Data}
(the latter referred to the year of Varicella infection), it adds
rows for each combination of \code{Id_Medico} and \code{nPaz} reporting the
status of "they was infected" for all the years from 2004 to 2014
in a column named \code{is_was_positive}
}
\examples{
db <- tibble::tribble(
  ~ Id_Medico, ~nPaz, ~Data, ~other, ~anno,
  1, 1, lubridate::ymd("2005-12-31"), "foo", 2005L,
  2, 1, lubridate::ymd("2005-12-31"), "foo", 2005L,
  1, 2, lubridate::ymd("2008-12-31"), "foo", 2008L,
  3, 3, lubridate::ymd("2007-12-31"), "foo", 2007L
)

expand_prepost_positive(db)

}
