% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_path.R
\name{get_data_path}
\alias{get_data_path}
\title{Path to data}
\usage{
get_data_path()
}
\value{
(chr) the path to the data folder
}
\description{
Path to data
}
\examples{
\dontrun{
  popolazione <- file.path(data_path, "popolazione.rda")
  popolazione
}
}
