% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_and_complete.R
\name{merge_and_complete}
\alias{merge_and_complete}
\title{Merge information and complete missing ones}
\usage{
merge_and_complete(main_df, gold, popolazione)
}
\arguments{
\item{popolazione}{}
}
\value{

}
\description{
Merge main varicella db with the gold standard and the information
contained inside popolazione. Moreover, complete the missing
\code{class}es (in the original db that are not ids in the gold standard)
with "negative" label, for each year
}
