% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_records.R
\name{merge_records}
\alias{merge_records}
\title{Merge patient's records}
\usage{
merge_records(db, meta_vars)
}
\arguments{
\item{meta_vars}{(chr) included variables in \code{x} that should not be
merged as a text. The will be summarized as a single list-entry in
the final one-row dataframe}

\item{x}{(data frame) dataframe of data and metadata for a single
ID, for a single year (and a set, and all other supposed unique
variables for that ID-year group)}
}
\value{
a dataframe with a single row
}
\description{
Merge together all the record from a patient's ID (in a year), by
separating each field by an "\strong{EOF}" token, and each observation
by an "\strong{EOR}" one.
}
\examples{

db <- tibble::tribble(
  ~txt_1, ~txt_2, ~meta_txt, ~ meta_num,
   "faa tra", "bar", "tar", 12,
   "fee fir", NA, "ter", NA,
   NA, "bir", "tir", 17,
   "foo", "bor", NA, 24,
   "fuu", "bur", "tur", 82
) |>
  dplyr::mutate(date = sample(1:5 + Sys.Date()))

meta_vars <- c("meta_txt", "meta_num", "date")

merge_records(db, meta_vars) |>
  dplyr::glimpse()


}
