% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_text_meta_records.R
\name{merge_text_meta_records}
\alias{merge_text_meta_records}
\title{Merge records by id and year}
\usage{
merge_text_meta_records(db, meta_vars = character())
}
\arguments{
\item{db}{data frame}
}
\value{
a data frame with a single row for each combination of
id-year
}
\description{
Merge textual and meta information for each id by year.
\itemize{
\item All unique, non textual, information are kept as they are
\item Each meta information is merged in a single row, concatenated in
a global column-list (in the final data frame.
\item All textual information are merged in a single row and in a single
column. Each field separated by a "\strong{EOF}" token, each record
separated by a "\strong{EOR}" token.
}
}
\details{
\code{db} must have column \code{set}
("train", "validation", adn possibly "test"); \code{anno} (the year
considered); \code{n_paz} (patient id); \code{class} ("positive", "negative")
}
\examples{


db <- tibble::tribble(
  ~id_medico, ~n_paz, ~anno, ~class, ~set, ~txt_1, ~txt_2, ~meta_txt, ~ meta_num,
   1, 1, 2005, "positive", "train", "faa tra", "bar", "tar", 12,
   2, 1, 2006, "negative", "validation", "fee fir", NA, "ter", NA,
   1, 2, 2008, "negative", "train", NA, "bir", "tir", 17,
   1, 3, 2006, "positive", "train", "foo", "bor", NA, 24,

   # patient 4 has two id_medico qithin same year and set with
   # different class. "positive class" will predominate
   3, 4, 2004, "positive", "train", "fuu", "bur", "tur", 82,
   4, 4, 2004, "negative", "train", "fuu", "bur", "tur", 82
) |>
  dplyr::mutate(date = rev(1:6 + Sys.Date()))

meta_vars <- c("id_medico", "meta_txt", "meta_num", "date")

res <- merge_text_meta_records(db, meta_vars)
res

}
