% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{pedianet_fasttext}
\alias{pedianet_fasttext}
\title{FastText pretrained dictionaries for PediNET}
\description{
Textual, single-space separated for 122607 words. The
first line report the dimensions (words and vector length). Each
further line start with a word and its floating-point vector
rapresentation
}
\details{
\itemize{
\item \strong{pedianet_fasttext_100.vec}: A dictionary containing the
skipgram 100-legnth dense vectors representation for the words
in the PediaNET database.
\item \strong{pedianet_fasttext_300.vec}: A dictionary containing the
skipgram 300-legnth dense vectors representation for the words
in the PediaNET database.
}
}
