% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_varic.R
\name{read_varic}
\alias{read_varic}
\title{Read varicella2 DBs}
\usage{
read_varic(path, db = c(NULL, "popolazione", "positive_gold", "varic_df"))
}
\arguments{
\item{path}{(chr) path to RDS to import}

\item{db}{(chr) one of  "popolazione", "positive_gold", or
"varic_df". Note that this is a mandatory field.}
}
\value{
a \link[tibble:tibble-package]{tibble} for the corresponding
\code{path} dataset.
}
\description{
Read varicella2 DBs
}
\examples{
\dontrun{
  read_varic(get_data_path(), "popolazione")
  read_varic(get_data_path(), "positive_gold")
  read_varic(get_data_path(), "varic_df")
}

}
