% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varicella.due-package.R
\docType{package}
\name{varicella.due-package}
\alias{varicella.due}
\alias{varicella.due-package}
\title{varicella.due: Varicella Due}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Corrado Lanera \email{corrado.lanera@ubep.unipd.it} (\href{https://orcid.org/0000-0002-0520-7428}{ORCID})

Other contributors:
\itemize{
  \item Andrea Francavilla \email{andrea.francavilla@ubep.unipd.it} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID}) [contributor]
  \item Ileana Baldi \email{ileana.baldi@ubep.unipd.it} (\href{https://orcid.org/0000-0002-8578-9164}{ORCID}) [contributor]
  \item UBEP [copyright holder]
}

}
\keyword{internal}
