% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_varicella_mixdb.R
\name{create_varicella_mixdb}
\alias{create_varicella_mixdb}
\title{Create the main mixdb for the vericella project}
\usage{
create_varicella_mixdb(data_path = get_data_path(), gold)
}
\arguments{
\item{data_path}{(chr) path to the data folder}

\item{gold}{(chr) pedinet gold standard}
}
\value{
invisibly `TRUE``
}
\description{
Create the main \link{mixdb} for the varicella project
}
