% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_prepost_positive.R
\name{expand_prepost_positive}
\alias{expand_prepost_positive}
\title{Expand previous and post positive status}
\usage{
expand_prepost_positive(db_gold, pop)
}
\arguments{
\item{db_gold}{(data frame) positive database with \code{id_medico}, \code{n_paz},
and \code{data} as the minimum set of columns. Column \code{data} reporting the
year of the Varicella infection for the corresponding patient.}

\item{pop}{(data frame) database with all population IDs stored in a
\code{n_paz} variable with corresponding \code{id_medico} for each one.}
}
\value{
the original \link[tibble:tibble-package]{tibble} with the
\code{class} column added as last one.
}
\description{
Given a dataset with columns \code{id_medico}, \code{n_paz}, and \code{data}
(the latter referred to the year of Varicella infection), it adds
rows for each combination of \code{id_medico} and \code{n_paz} reporting the
status of "they was infected" for all the years from 2004 to 2014
in a column named \code{class}
}
\examples{
db_gold <- tibble::tribble(
  ~n_paz, ~anno, ~class,
  1, 2005L, "positive",
  2, 2008L, "positive",
  3, 2007L, "positive"
)

pop <- data.frame(
  id_medico = c(1, 2, 1, 3, 3),
  n_paz = c(1, 1, 2, 3, 4),
  foo = "bar"
)

expand_prepost_positive(db_gold, pop)

}
