% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_id_records.R
\name{merge_text_meta_records}
\alias{merge_text_meta_records}
\title{Merge records by id and year}
\usage{
merge_text_meta_records(db, meta_vars = character())
}
\arguments{
\item{db}{data frame}
}
\value{

}
\description{
Merge textual and meta information for each id by year.
\itemize{
\item All unique, non textual, information are kept as they are
\item Each meta information is merged in a single row, concatenated in
a global column-list (in the final data frame.
\item All textual information are merged in a single row and in a single
column. Each field separated by a "\strong{EOF}" token, each record
separated by a "\strong{EOR}" token.
}
}
\details{
\code{db} must have column \code{set}
("train", "validation", adn possibly "test"); \code{anno} (the year
considered); \code{n_paz} (patient id); \code{class} ("positive", "negative")
}
