% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_varic.R
\name{read_varic}
\alias{read_varic}
\title{Read varicella2 DBs}
\usage{
read_varic(path, db = c("popolazione", "positive_gold", "varic_df"))
}
\arguments{
\item{what}{(chr) one of "popolazione", "positive_gold", "varic_df"}
}
\value{
a \link[tibble:tibble-package]{tibble} for the corresponding
\link{what} dataset.
}
\description{
Read varicella2 DBs
}
\examples{
read_varic("popolazione")
read_varic("positive_gold")
read_varic("varic_df")

}
